Update delimeter to be pipes to allow for commmas in messages
With a traditional .csv file, commas are the delimeters separating each
datum from each other. If we want to include messages, it is very likely
that there will be commas in it, and that would throw off the simple
parsing logic here. Therefore, we opted for pipe separated values (.psv)
because They will almost never be typed. In the rare case that they would
be, then the program will have to replace it with a '^' instead, but that
is for the next commit.