Fix OSError when redirecting output in terminal
In the terminal, when redirecting output into a file (and causing
output_everything to be invoked), the call "os.get_terminal_size()"
raised an OSError, presumably because the file it was being redirected
to is not a tty and can't give that kind of information. Updated to
have a try/except statement for OSError when calling "os.get_terminal_size()"
and to assign number of terminal lines to positive infinity if OSError is raised.